Fix more test errors.  Index root item by user.uid instead of user.id since user.id is always -1.
git-svn-id: http://svn.osafoundation.org/server/cosmo/branches/0.6.1@3726 5fd0b11a-f2f7-0310-948e-c8f80ebed640

+++ cosmo/src/main/java/org/osaf/cosmo/service/impl/StandardUserService.java
@@ -422,6 +422,8 @@ public class StandardUserService implements UserService {
      * collection of all the user's items.
      */
     private void removeUserAndItems(User user) {
+        if(user==null)
+            return;
         HomeCollectionItem home = contentDao.getRootItem(user);
         contentDao.removeCollection(home);
         userDao.removeUser(user);
